graphQlDjango/firstproject/heros/requirements.txt
Django==2.2.3
gunicorn[gevent]==19.9.0
# django-cors-headers
# djangorestframework==3.10.1
# requests==2.2.0
# xhtml2pdf==0.2.3
# matplotlib==3.1.1
# numpy==1.16.2
# pandas==0.25.0
# django-rest-swagger==2.1.1
# packaging==19.1
# drf-yasg==1.16.1
# boto3==1.9.246
# django-storages==1.7.2
graphene-django==2.6.0
django-graphiql==0.4.4
django-graphql-jwt==0.2.2